requirements.txt
fastapi
uvicorn
pyjwt
bcrypt
pydantic
pydantic-settings
redis
sqlalchemy
asyncpg
alembic
boto3
httpx
email-validator
python-dotenv
python-multipart
